ui: publish distribution settings pages
