+++ .changeset/distribution-settings-pages.md
@@ -1,0 +1,5 @@
+---
+"@voyantjs/distribution-ui": minor
+---
+
+Publish the channels settings and channel sync page compositions with package data hooks and app-provided API client extension points.
