+++ packages/distribution-ui/src/i18n/en.ts
@@ -329,6 +329,41 @@ export const distributionUiEn = {
       },
     },
   },
+  settings: {
+    channelsPage: {
+      title: "Channels",
+      description:
+        "Define where your products are sold: direct, OTA, reseller, and marketplace channels.",
+      addChannel: "Add Channel",
+      empty:
+        "No channels yet. Create channels like Website, Mobile App, or Viator to control where products are sold.",
+      edit: "Edit",
+      delete: "Delete",
+      deleteConfirm: "Delete this channel?",
+      deleteDescription: "Products assigned to this channel will be unlinked.",
+      editSheetTitle: "Edit Channel",
+      newSheetTitle: "New Channel",
+      nameLabel: "Name",
+      namePlaceholder: "Website",
+      kindLabel: "Kind",
+      statusLabel: "Status",
+      websiteLabel: "Website",
+      websitePlaceholder: "https://partner.example.com",
+      primaryContactLabel: "Primary Contact",
+      primaryContactPlaceholder: "Jane Doe",
+      contactEmailLabel: "Contact Email",
+      contactEmailPlaceholder: "partner@example.com",
+      saveChanges: "Save Changes",
+      createChannel: "Create Channel",
+      validationNameRequired: "Name is required",
+      validationInvalidUrl: "Must be a valid URL",
+      validationInvalidEmail: "Must be a valid email",
+      paginationShowing: "Showing {count} of {total}",
+      paginationPage: "Page {page} / {pageCount}",
+      paginationPrevious: "Previous",
+      paginationNext: "Next",
+    },
+  },
   details: {
     channel: {
       notFound: "Channel not found",
