+++ packages/distribution-ui/src/i18n/ro.ts
@@ -333,6 +333,41 @@ export const distributionUiRo = {
       },
     },
   },
+  settings: {
+    channelsPage: {
+      title: "Canale",
+      description:
+        "Defineste unde sunt vandute produsele tale: canale directe, OTA, reseller si marketplace.",
+      addChannel: "Adauga canal",
+      empty:
+        "Nu exista canale momentan. Creeaza canale precum Website, Mobile App sau Viator pentru a controla unde sunt vandute produsele.",
+      edit: "Editeaza",
+      delete: "Sterge",
+      deleteConfirm: "Stergi acest canal?",
+      deleteDescription: "Produsele alocate acestui canal vor fi deconectate.",
+      editSheetTitle: "Editeaza canal",
+      newSheetTitle: "Canal nou",
+      nameLabel: "Nume",
+      namePlaceholder: "Website",
+      kindLabel: "Tip",
+      statusLabel: "Status",
+      websiteLabel: "Website",
+      websitePlaceholder: "https://partner.example.com",
+      primaryContactLabel: "Contact principal",
+      primaryContactPlaceholder: "Jane Doe",
+      contactEmailLabel: "Email contact",
+      contactEmailPlaceholder: "partner@example.com",
+      saveChanges: "Salveaza modificarile",
+      createChannel: "Creeaza canal",
+      validationNameRequired: "Numele este obligatoriu",
+      validationInvalidUrl: "Trebuie sa fie un URL valid",
+      validationInvalidEmail: "Trebuie sa fie un email valid",
+      paginationShowing: "Afisare {count} din {total}",
+      paginationPage: "Pagina {page} / {pageCount}",
+      paginationPrevious: "Anterior",
+      paginationNext: "Urmator",
+    },
+  },
   details: {
     channel: {
       notFound: "Canalul nu a fost gasit",
